update: convert the organizations to prisma

+++ organizations/organizations.controllers copy.js
@@ -1,21 +1,21 @@
-import { EmployeeRole, PrismaClient } from "@prisma/client";
+import { PrismaClient } from "@prisma/client";
 import asyncHandler from "express-async-handler";
-import ShortUniqueId from "short-unique-id";
+import Joi from "joi";
 
 const prisma = new PrismaClient();
+// TODO: Create transfer organization to another user endpoints 
 
-const { randomUUID } = new ShortUniqueId({ length: 6 });
 
-// Create a new organization
-const createOrganization = asyncHandler(async (req, res) => {
-  const { name } = req.body;
-  const { id: userId, email } = req.user;
+// Create a new organizations
+const createOrganization = async (req, res) => {
+  const { id: createdById } = req.user;
+  const { name, address, logo } = req.body;
 
   // Convert the organization name to lowercase
   const lowercaseName = name.toLowerCase();
 
   try {
-    // Check if an organization with the same name already exists
+     // Check if an organization with the same name already exists
     const existingOrganization = await prisma.organization.findFirst({
       where: { name: lowercaseName },
     });
@@ -23,7 +23,7 @@ const createOrganization = asyncHandler(async (req, res) => {
     if (existingOrganization) {
       return res
         .status(400)
-        .json({ error: `Organization name ${lowercaseName} already exists.` });
+        .json({ error: `Organization name ${name} already exists.` });
     }
 
     // Check if a user with the provided email already exists
@@ -34,20 +34,21 @@ const createOrganization = asyncHandler(async (req, res) => {
     // Create a new organization with the provided data
     const newOrganization = await prisma.organization.create({
       data: {
-        name: lowercaseName,
-        imageUrl: "img-1",
-        inviteCode: randomUUID(),
-        userId,
-      },
-    });
-
-    // Create a new employee associated with the user who created the organization
-    const newEmployee = await prisma.employee.create({
-      data: {
-        email: email,
-        role: EmployeeRole.OWNER,
-        userId,
-        organizationId: newOrganization.id,
+        name: name,
+        address: address || null,
+        logo: logo || null,
+        owner: {
+          connect: { id: createdById },
+        },
+        employees: {
+          create: [
+            {
+              fullName: existingUser.fullName,
+              email: existingUser.email,
+              role: "Owner",
+            },
+          ],
+        },
       },
     });
 
@@ -55,15 +56,18 @@ const createOrganization = asyncHandler(async (req, res) => {
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
-});
+};
 
 // Get all organizations
 const getAllOrganizations = asyncHandler(async (req, res) => {
-  const { email, id } = req.user;
+  const { email } = req.user;
   try {
     const organizations = await prisma.organization.findMany({
       where: {
-        OR: [{ userId: id }, { employees: { some: { email } } }],
+        OR: [
+          { createdById: req.user.id },
+          { employees: { some: { email: email } } },
+        ],
       },
       include: {
         employees: {
@@ -76,9 +80,7 @@ const getAllOrganizations = asyncHandler(async (req, res) => {
 
     // Check if the user is not part of any organizations
     if (organizations.length === 0) {
-      return res
-        .status(200)
-        .json({ message: "You're not a part of any organization" });
+      return res.status(200).json({message: "You're not a part of any organization"}); 
     }
     res.status(200).json(organizations);
   } catch (error) {
@@ -86,14 +88,15 @@ const getAllOrganizations = asyncHandler(async (req, res) => {
   }
 });
 
-// Get an organization by ID
+// Get an Organization by ID
 const getOrganizationById = asyncHandler(async (req, res) => {
   const { id } = req.params;
 
   try {
     const organization = await prisma.organization.findUnique({
       where: { id },
-      include: {
+      select: {
+        id: true,
         employees: {
           select: {
             id: true,
@@ -103,7 +106,15 @@ const getOrganizationById = asyncHandler(async (req, res) => {
     });
 
     if (!organization) {
-      res.status(404).json(`Organization ${id} not found`);
+      return res.status(404).json(`Organization ${id} not found`);
+    }
+
+    // Check if the user is authorized to access this organization
+    if (
+      organization.createdById !== req.user.id &&
+      !organization.employees.some((employee) => employee.id === req.user.id)
+    ) {
+      return res.status(403).json("Unauthorized to access this organization");
     }
 
     res.status(200).json(organization);
@@ -111,23 +122,25 @@ const getOrganizationById = asyncHandler(async (req, res) => {
     res.status(500).json({ error: error.message });
   }
 });
-// Update an organization by ID
+
+// Update an Organization by ID
 const updateOrganization = asyncHandler(async (req, res) => {
   const { id } = req.params;
 
   try {
-    // Check if the organization with the specified ID exists
-    const existingOrganization = await prisma.organization.findUnique({
+    const organization = await prisma.organization.findUnique({
       where: { id },
     });
 
-    if (!existingOrganization) {
-      return res
-        .status(404)
-        .json({ error: `Organization with ID ${id} not found.` });
+    if (!organization) {
+      return res.status(404).json(`Organization ${id} not found`);
     }
 
-    // Update the organization data
+    // Check if the user is authorized to update the organization (e.g., user is the owner)
+    if (organization.createdById !== req.user.id) {
+      return res.status(403).json("Unauthorized to update this organization");
+    }
+
     const updatedOrganization = await prisma.organization.update({
       where: { id },
       data: req.body,
@@ -144,13 +157,26 @@ const deleteOrganization = asyncHandler(async (req, res) => {
   const { id } = req.params;
 
   try {
+    const organization = await prisma.organization.findUnique({
+      where: { id },
+    });
+
+    if (!organization) {
+      return res.status(404).json(`Organization ${id} not found`);
+    }
+
+    // Check if the user is authorized to delete the organization (e.g., user is the owner)
+    if (organization.createdById !== req.user.id) {
+      return res.status(403).json("Unauthorized to delete this organization");
+    }
+
     await prisma.organization.delete({
       where: { id },
     });
 
-    res.status(204).json(`Organization ${id} deleted`);
+    return res.status(204).json(`Organization ${id} deleted`);
   } catch (error) {
-    res.status(500).json({ error: error.message });
+    return res.status(500).json({ error: error.message });
   }
 });
